Finalize taskmanager for workshop

+++ public/js/AppExampleData.js
@@ -1,13 +1,27 @@
 module.exports = {
     init: function() {
-        localStorage.clear();
-        localStorage.setItem('taskCollection', JSON.stringify([
-            {
-                id: 'task1',
-                name: 'First Task',
-                text: 'Create this new visibility report for customer XYZ',
-                assignee: 'Martin'
-            }
-        ]));
+        if (!localStorage.getItem('taskCollection')) {
+            localStorage.clear();
+            localStorage.setItem('taskCollection', JSON.stringify([
+                {
+                    id: 'dummytask1',
+                    title: 'First Task',
+                    description: 'Create this new visibility report for customer XYZ',
+                    completed: false
+                },
+                {
+                    id: 'dummytask2',
+                    title: 'Second Task',
+                    description: 'Check out the second default task',
+                    completed: false
+                },
+                {
+                    id: 'dummytask3',
+                    title: 'Another Task',
+                    description: '',
+                    completed: false
+                }
+            ]));
+        }
     }
 };
